Move package com.catenax to net.catenax
