+++ coreservices/partsrelationshipservice/broker-proxy/src/main/java/net/catenax/brokerproxy/controllers/ApiErrorsConstants.java
@@ -10,7 +10,7 @@
 package net.catenax.brokerproxy.controllers;
 
 
-import com.catenax.partsrelationshipservice.annotations.ExcludeFromCodeCoverageGeneratedReport;
+import net.catenax.prs.annotations.ExcludeFromCodeCoverageGeneratedReport;
 
 /**
  * Helper class contains all Api Errors as reusable constants.
