+++ coreservices/partsrelationshipservice/broker-proxy/src/main/java/net/catenax/brokerproxy/controllers/BrokerProxyExceptionHandler.java
@@ -9,10 +9,10 @@
 //
 package net.catenax.brokerproxy.controllers;
 
-import com.catenax.partsrelationshipservice.annotations.ExcludeFromCodeCoverageGeneratedReport;
-import com.catenax.partsrelationshipservice.dtos.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import net.catenax.brokerproxy.exceptions.MessageProducerFailedException;
+import net.catenax.prs.annotations.ExcludeFromCodeCoverageGeneratedReport;
+import net.catenax.prs.dtos.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindException;
