+++ coreservices/partsrelationshipservice/integration-tests/src/test/java/net/catenax/prs/smoketest/SmokeTests.java
@@ -17,8 +17,8 @@ import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import static com.catenax.partsrelationshipservice.dtos.PartsTreeView.AS_MAINTAINED;
 import static io.restassured.RestAssured.given;
+import static net.catenax.prs.dtos.PartsTreeView.AS_MAINTAINED;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.hasSize;
