+++ coreservices/partsrelationshipservice/prs-api/src/main/java/net/catenax/prs/services/EventMessageRouter.java
@@ -9,9 +9,9 @@
 //
 package net.catenax.prs.services;
 
-import com.catenax.partsrelationshipservice.dtos.events.PartRelationshipsUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import net.catenax.prs.dtos.events.PartRelationshipsUpdateRequest;
 import org.springframework.kafka.annotation.KafkaHandler;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.KafkaHeaders;
