+++ coreservices/partsrelationshipservice/prs-api/src/test/java/net/catenax/prs/requests/RequestMother.java
@@ -1,7 +1,7 @@
 package net.catenax.prs.requests;
 
-import com.catenax.partsrelationshipservice.dtos.PartsTreeView;
 import com.github.javafaker.Faker;
+import net.catenax.prs.dtos.PartsTreeView;
 import net.catenax.prs.entities.EntitiesMother;
 import net.catenax.prs.entities.PartIdEntityPart;
