+++ coreservices/partsrelationshipservice/prs-models/src/main/java/net/catenax/prs/annotations/ValueOfEnum.java
@@ -7,10 +7,10 @@
 // See the LICENSE file(s) distributed with this work for
 // additional information regarding license terms.
 //
-package com.catenax.partsrelationshipservice.annotations;
+package net.catenax.prs.annotations;
 
 
-import com.catenax.partsrelationshipservice.validators.ValueOfEnumValidator;
+import net.catenax.prs.validators.ValueOfEnumValidator;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
