+++ coreservices/partsrelationshipservice/prs-models/src/main/java/net/catenax/prs/dtos/PartId.java
@@ -7,7 +7,7 @@
 // See the LICENSE file(s) distributed with this work for
 // additional information regarding license terms.
 //
-package com.catenax.partsrelationshipservice.dtos;
+package net.catenax.prs.dtos;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -15,20 +15,19 @@ import lombok.Builder;
 import lombok.Value;
 
 import javax.validation.constraints.NotBlank;
-import java.net.URL;
 
-/*** API type for aspect name/url entry. */
-@Schema(description = "Aspect location data")
+/*** API type for a part identifier. */
+@Schema(description = "Unique part identifier")
 @Value
 @Builder(toBuilder = true, setterPrefix = "with")
-@JsonDeserialize(builder = Aspect.AspectBuilder.class)
+@JsonDeserialize(builder = PartId.PartIdBuilder.class)
 @SuppressWarnings("PMD.CommentRequired")
-public class Aspect {
+public class PartId {
     @NotBlank
-    @Schema(description = "Aspect name", example = "CE")
-    private String name;
+    @Schema(description = "Readable ID of manufacturer including plant")
+    private String oneIDManufacturer;
 
     @NotBlank
-    @Schema(description = "URL location of aspect data", example = "http://aspects-url/CE", implementation = URL.class)
-    private String url;
+    @Schema(description = "Unique identifier of a single, unique physical (sub)component/part/batch, given by its manufacturer. For a vehicle, the Vehicle Identification Number (VIN).")
+    private String objectIDManufacturer;
 }
