+++ coreservices/partsrelationshipservice/prs-models/src/main/java/net/catenax/prs/dtos/PartInfo.java
@@ -7,7 +7,7 @@
 // See the LICENSE file(s) distributed with this work for
 // additional information regarding license terms.
 //
-package com.catenax.partsrelationshipservice.dtos;
+package net.catenax.prs.dtos;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -15,19 +15,22 @@ import lombok.Builder;
 import lombok.Value;
 
 import javax.validation.constraints.NotNull;
+import java.util.List;
 
-/*** API type for a relationship between two parts. */
-@Schema(description = "Link between two parts.")
+/*** API type for part information retrieved alongside a parts tree. */
+@Schema(description = "Information about parts")
 @Value
 @Builder(toBuilder = true, setterPrefix = "with")
-@JsonDeserialize(builder = PartRelationship.PartRelationshipBuilder.class)
+@JsonDeserialize(builder = PartInfo.PartInfoBuilder.class)
 @SuppressWarnings("PMD.CommentRequired")
-public class PartRelationship {
+public class PartInfo {
     @NotNull
-    @Schema(description = "Unique part identifier of the parent in the relationship.")
-    private PartId parent;
+    @Schema(description = "The part identifier.")
+    private PartId part;
 
-    @NotNull
-    @Schema(description = "Unique part identifier of the child in the relationship.")
-    private PartId child;
+    @Schema(description = "Type of material, (sub)component/part or vehicle. May not be set.", example = "gearbox")
+    private String partTypeName;
+
+    @Schema(description = "List of aspect locations.")
+    private List<Aspect> aspects;
 }
