+++ coreservices/partsrelationshipservice/prs-models/src/main/java/net/catenax/prs/dtos/PartRelationshipsWithInfos.java
@@ -7,7 +7,7 @@
 // See the LICENSE file(s) distributed with this work for
 // additional information regarding license terms.
 //
-package com.catenax.partsrelationshipservice.dtos;
+package net.catenax.prs.dtos;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.swagger.v3.oas.annotations.media.Schema;
